Test Serialize & Deserialize
